+++ MedV4D/include/MedV4D/Imaging/filters/ConvolutionFilter.h
@@ -30,7 +30,7 @@ class ConvolutionFilter2D
 	: public AImage2DFilter< ImageType, ImageType >
 {
 public:	
-	static const unsigned Dimension = ImageTraits< ImageType >::Dimension;
+	static const size_t Dimension = ImageTraits< ImageType >::Dimension;
 	typedef typename ImageTraits< ImageType >::ElementType ElementType;
 	typedef AImage2DFilter< ImageType, ImageType > PredecessorType;
 	typedef typename ConvolutionMask<2,MatrixElement>::Ptr	MaskPtr;
